Add AuthProvider tests; rename tests to .tsx
Rename auth tests from .ts to .tsx and add React Testing Library imports. Add AuthProvider test suite in both admin-app and referral-app covering: rendering children after auth initializes, showing fallback while auth is pending, and ensuring init is only called once across mounts. Also add a keycloak export assertion in referral-app to verify the instance is exported.
